Adressen laten cascaden vanaf Klant om transientobjectexception te voorkomen. Klant object kan nu via hibernate in database worden opgeslagen. Test werkt helaas nog niet, vanwege ander probleem wat in comment in KlantServiceTest wordt beschreven.

+++ floorinxs-crm/src/main/java/org/workshop2/floorinxs/dao/KlantDaoImpl.java
@@ -5,6 +5,6 @@ import org.springframework.transaction.annotation.Transactional;
 import org.workshop2.floorinxs.entity.Klant;
 
 @Repository
-public class KlantDaoImpl extends AbstractDao<Klant, Integer> implements KlantDao {
+public class KlantDaoImpl extends AbstractDao<Klant, Long> implements KlantDao {
     
 }
