+++ floorinxs-crm/src/main/java/org/workshop2/floorinxs/service/KlantServiceImpl.java
@@ -4,6 +4,7 @@ import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.workshop2.floorinxs.dao.AdresDao;
 import org.workshop2.floorinxs.dao.KlantDao;
 import org.workshop2.floorinxs.entity.Klant;
 
@@ -25,7 +26,7 @@ public class KlantServiceImpl implements KlantService {
     }
 
     @Override
-    public Klant findById(int id) {
+    public Klant findById(long id) {
         return klantDao.readById(id);
     }
